test: update ceo-inbox-update-folders tests for name->ID call order (#1216)
The pre-existing tests/unit/skills/ceo-inbox-update-folders.test.ts used a
fixed mockResolvedValueOnce sequence (getMessage -> PUT). The handler now
calls listFolders first to resolve label names to IDs, so the sequence
misaligned and 4 tests failed in CI.

Rewrite that suite to route the fetch mock by URL+method (robust to call
order), preserve all prior cases (validation, add/remove, empty-folders
fallback, dedup, getMessage-throws), and fold in the new name->ID /
create / removed_unresolved / error-detail cases. Delete the duplicate
co-located skills/ceo-inbox-update-folders/handler.test.ts.
